added kinds of precedence graphs
