+++ src/fmmlx_mlm_structure/attribute_precedence_graph.py
@@ -1,0 +1,23 @@
+from src.fmmlx_mlm_structure.precedence_graph import PropertyPrecedenceGraph
+
+
+class AttributePrecedenceGraph(PropertyPrecedenceGraph):
+
+    def __init__(self):
+        super().__init__()
+        self.max_level: int = 0
+
+    def set_inst_levels_for_attributes(self):
+        inst_level: int = 0
+        for pg_list in self.get_topological_generations():
+            for pg in pg_list:
+                for attr in pg.get_model_properties():
+                    attr.set_proposed_inst_level(inst_level)
+            inst_level += 1
+        self.max_level = inst_level - 1
+
+    def has_deepening_potential(self) -> bool:
+        return self.max_level > 0
+
+    def get_max_level(self) -> int:
+        return self.max_level
