+++ src/precedence_analysis_tester.py
@@ -25,8 +25,8 @@ class PrecedenceAnalysisTester(Tester):
                 self.init_test("supermarket_sales.csv", csv_columns=selected_columns_supermarket)
             case _:
                 raise Exception("Invalid test variant (variant number: " + str(self.variant) + ") unspecified")
-        self.md_instance.perform_property_precedence_analysis(print_attribute_relations=False,
-                                                              print_slot_comparisons=False)
+        self.md_instance.perform_property_precedence_analysis(print_attribute_relations=True,
+                                                              print_slot_comparisons=True)
         if self.export_model:
             self.export_model_xml()
